Refactor generate into a folder
