+++ src/tensora/expression/__init__.py
@@ -1,3 +1,2 @@
-from ._deparse_to_taco import deparse_to_taco
 from ._exceptions import InconsistentDimensionsError, MutatingAssignmentError
 from ._parser import parse_assignment
