+++ src/tensora/generate/__init__.py
@@ -1,0 +1,2 @@
+from ._base import TensorCompiler, generate_c_code
+from ._taco import TacoError
